# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank_REST
  description: backend-приложение на Java (Spring Boot) для управления банковскими
    картами
  contact:
    name: API support
    url: https://github.com/22crystyle
    email: shimorowm@gmail.com
  version: 1.0.0
servers:
- url: http://localhost:1024
  description: Local server
...
paths:
...
  /api/v1/accounts:
    get:
      tags:
      - Accounts
      summary: Get paginated list of accounts
      description: "Returns a paginated list of account resources, requires ADMIN\
        \ role."
      operationId: getAccounts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
      responses:
        "200":
          description: A page of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAccountResponse"
        "400":
          description: The parameters do not match the validation
    post:
      tags:
      - Accounts
      summary: Create new account
      description: "Receives account data for a new user, saves it to the database,\
        \ and returns the created object with status code 201 and a Location header."
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
        required: true
      responses:
        "201":
          description: Account successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "401":
          description: Unauthorized
        "403":
          description: Access Denied
        "409":
          description: Conflict. User with unique keys already exists
...
  /api/v1/accounts/{id}:
    get:
      tags:
      - Accounts
      summary: Get account by ID
      description: Returns account details for the specified account ID
      operationId: getAccountById
      parameters:
      - name: id
        in: path
        description: ID of the account to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "403":
          description: Access Denied
        "404":
          description: Account not found
    delete:
      tags:
      - Accounts
      summary: Delete account by id
      description: Deletes the account with the specified ID. Requires ADMIN role.
      operationId: deleteAccount
      parameters:
      - name: id
        in: path
        description: ID of the account to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Account is successfully deleted
        "404":
          description: Account not found
  /api/v1/accounts/{id}/cards:
    get:
      tags:
      - Accounts
      summary: Get all cards for an account
      description: Returns a list of cards associated with the specified account ID.
      operationId: getAccountCardsByUserId
      parameters:
      - name: id
        in: path
        description: ID of the account
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
components:
  schemas:
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        owner:
          $ref: "#/components/schemas/AccountResponse"
        status:
          $ref: "#/components/schemas/CardStatusResponse"
        balance:
          type: number
    CardStatusResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
        name:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          minLength: 1
        description:
          type: string
...
    AccountRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        roleId:
          type: integer
          format: int32
...
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        offset:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
    PageAccountResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        last:
          type: boolean
        totalElements:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
  securitySchemes:
    Basic:
      type: http
      scheme: basic